+++ cli/dbedit/cli.py
@@ -9,11 +9,28 @@ def list_():
     pass
 
 
+#### $ cidc admin list clinical ####
+@click.command("clinical")
+@click.argument("trial_id", required=True, type=str)
+def list_clinical(trial_id: str):
+    """
+    List clinical files for a given trial
+
+    TRIAL_ID is the id of the trial to affect
+    """
+    core.connect()
+    list.list_clinical(trial_id=trial_id)
+
+
 #### $ cidc admin list shipments ####
 @click.command("shipments")
 @click.argument("trial_id", required=True, type=str)
 def list_shipments(trial_id: str):
-    """Listing shipments for a given trial"""
+    """
+    List shipments for a given trial
+
+    TRIAL_ID is the id of the trial to affect
+    """
     core.connect()
     list.list_shipments(trial_id=trial_id)
 
@@ -25,12 +42,35 @@ def remove_():
     pass
 
 
+#### $ cidc admin remove clinical ####
+@click.command("clinical")
+@click.argument("trial_id", required=True, type=str)
+@click.argument("target_id", required=True, type=str)
+def remove_clinical(trial_id: str, target_id: str):
+    """
+    Remove a given clinical data file from a given trial's metadata
+    as well as remove the file itself from the portal.
+
+    TRIAL_ID is the id of the trial to affect
+    TARGET_ID is the object_url of the clinical data to remove
+        not including {trial_id}/clinical/
+        special value * for all files for this trial
+    """
+    core.connect()
+    remove.remove_clinical(trial_id=trial_id, target_id=target_id)
+
+
 #### $ cidc admin remove shipment ####
 @click.command("shipment")
 @click.argument("trial_id", required=True, type=str)
 @click.argument("target_id", required=True, type=str)
 def remove_shipment(trial_id: str, target_id: str):
-    """Deleting a given shipment from a given trial"""
+    """
+    Remove a given shipment from a given trial's metadata
+
+    TRIAL_ID is the id of the trial to affect
+    TARGET_ID is the manifest_id of the shipment to remove
+    """
     core.connect()
     remove.remove_shipment(trial_id=trial_id, target_id=target_id)