+++ cli/dbedit/core.py
@@ -17,7 +17,15 @@ Session = None
 DownloadableFiles, TrialMetadata, UploadJobs, Users = None, None, None, None
 
 
-def connect():
+def connect() -> None:
+    """
+    Set up this module to be able to make sqlalchemy calls
+    uses ENV as set by $ cidc config set-env
+    defaults to staging unless `prod` is specified ie no `dev` mode
+
+    If not already loaded, will ask for your database username
+    Asks for database password every time as to not store it
+    """
     ENV: str = get_env()
     if ENV not in ["prod", "staging"]:
         print("unknown ENV, applying to staging:", ENV)
@@ -76,7 +84,40 @@ def connect():
     Users = Base.classes.users
 
 
+def get_clinical_downloadable_files(
+    trial_id: str, *, session: Session
+) -> List[DownloadableFiles]:
+    """
+    Get all downloadable_files rows from clinical data files for the given trial
+
+    Parameters
+    ----------
+    trial_id: str
+        the id of the trial to affect
+    session: Session
+        a session created from this module's `Session` after `connect()` is called
+    """
+    return (
+        session.query(DownloadableFiles)
+        .filter(
+            DownloadableFiles.trial_id == trial_id,
+            DownloadableFiles.object_url.like("%/clinical/%"),
+        )
+        .all()
+    )
+
+
 def get_shipments(trial_id: str, *, session: Session) -> List[UploadJobs]:
+    """
+    Get all upload_jobs rows from successful manifest uploads for the given trial
+
+    Parameters
+    ----------
+    trial_id: str
+        the id of the trial to affect
+    session: Session
+        a session created from this module's `Session` after `connect()` is called
+    """
     # the "shipments" in trial_metadata don't have which
     # samples they came with, so we have to check the uploads
     shipments = []
@@ -92,3 +133,29 @@ def get_shipments(trial_id: str, *, session: Session) -> List[UploadJobs]:
         if "shipments" in u.metadata_patch:
             shipments.append(u)
     return shipments
+
+
+def get_trial_if_exists(
+    trial_id: str, *, with_for_update: bool = False, session: Session
+) -> TrialMetadata:
+    """
+    Get the trial_metadata row for the given trial for update
+    Exits with message if trial does not exist
+
+    Parameters
+    ----------
+    trial_id: str
+        the id of the trial to affect
+    session: Session
+        a session created from this module's `Session` after `connect()` is called
+    """
+    query = session.query(TrialMetadata).filter(TrialMetadata.trial_id == trial_id)
+    if with_for_update:
+        query = query.with_for_update()
+    trial: TrialMetadata = query.first()
+
+    if not trial:
+        print(f"Trial {trial_id} cannot be found")
+        exit()
+
+    return trial